requirements.txt
hvac==0.11.2
pyOpenSSL==22.0.0
termcolor==1.1.0
validators==0.20.0
